add start screen (prototype)
